R Code Best Practices
Update the below code such that it first checks if the “dplyr” package
is present. if it is already present, don’t install it just load the
package.If the package is not present, install it and then load it.

Change the below code so that the it doesn’t print package related
information while loading the plyr package.